\name{France2017P}
\alias{France2017P}
\docType{data}
\title{
2017 French Presidential Election. Regional provisional results.
}
\description{
Data frame containing the provisional results of the first and second rounds of the 2017 French presidential election in the 12 French continental regions (Auvergne-Rhone-Alpes, Bourgogne-Franche-Comte, Brittany, Centre-Val de Loire, Grand Est, Hauts-de-France, Ile-de-France, Normandy, Nouvelle-Aquitaine, Occitanie, Pays de la Loire, Provence-Alpes-Cote d'Azur) plus an additional region that covers Corsica and the rest of French overseas regions.
}
\usage{data(France2017P)}

\format{A table containing 13 observations and 12 variables:
\describe{
\item{ABSTENTION}{Number of people abstaining (NonVoters) and voting either blank
or null in the first-round of 2017 Presidential Election.}
\item{MACRON}{Number of votes gained at a national level for Emmanuel Macron in the
first-round of 2017 Presidential Election.}
\item{LE_PEN}{Number of votes gained at a national level for Marine Le Pen in the
first-round of 2017 Presidential Election.}
\item{FILLON}{Number of votes gained at a national level for Francois Fillon in the
first-round of 2017 Presidential Election.}
\item{MELENCHON}{Number of votes gained at a national level for Jean-Luc Melenchon in the
first-round of 2017 Presidential Election.}
\item{HAMON}{Number of votes gained at a national level for Benoit Hamon in the
first-round of 2017 Presidential Election.}
\item{DUPONT}{Number of votes gained at a national level for Nicolas Dupont-Aignan in the
first-round of 2017 Presidential Election.}
\item{OTHERS}{Number of votes gained at a national level for the rest of candidates in the
first-round of 2017 Presidential Election.}
\item{ABSTENTION2}{Number of people abstaining (NonVoters) in the second-round of 2017 Presidential Election.}
\item{BLANK_NULL}{Number of people voting either blank or null in the second-round of 2017 Presidential Election.}
\item{MACRON2}{Number of votes gained at a national level for Emmanuel Macron in the
second-round of 2017 Presidential Election}
\item{LE_PEN2}{Number of votes gained at a national level for Marine Le Pen in the
second-round of 2017 Presidential Election}
}}

\source{
Own elaboration from data available in \url{https://www.francetvinfo.fr/elections/resultats/}, retrieved 7 May 2017.
}

\keyword{datasets}
