% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.lphom}
\alias{print.summary.lphom}
\title{Print a summary of a lphom-family object}
\usage{
\method{print}{summary.lphom}(x, ..., margins = TRUE, digits = 2)
}
\arguments{
\item{x}{An \code{summary.lphom} class object.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{margins}{A \code{TRUE/FALSE} argument informing if the margins of the matrix should be displayed. Default \code{TRUE}.}

\item{digits}{Integer indicating the number of decimal places to be shown. Default, 2.}
}
\description{
Print method for \code{summary.lphom} objects
}
