% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lphom}
\alias{summary.lphom}
\title{Summarize a lphom-family object}
\usage{
\method{summary}{lphom}(object, ...)
}
\arguments{
\item{object}{An object output of a \strong{lphom} family algorithm.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
An object of class \code{"summary_lphom"}.
}
\description{
Summary method for objects obtained with an algorithm of the lphom-family (lphom, tslphom, nslphom, tslphom_dual, nslphom_joint, ....).
}
\examples{
mt.ns <- nslphom(France2017P[, 1:8] , France2017P[, 9:12], new_and_exit_voters= "raw")
summary(mt.ns)

}
\author{
Jose M. Pavia, \email{pavia@uv.es}
}
