\name{France2017D}
\alias{France2017D}
\docType{data}
\title{
2017 French Presidential Election. Department official results.
}
\description{
Data frame containing the official results of the first and second rounds of the 2017 French presidential
election in the 107 territorial French departments plus an artificial department that groups the
French electors living abroad.
}
\usage{data(France2017D)}

\format{A table containing 108 observations and 13 variables:
\describe{
\item{ABSTENTION}{Number of people abstaining (NonVoters) in the first-round of 2017 Presidential Election.}
\item{BLANK_NULL2}{Number of people voting either blank or null in the first-round of 2017 Presidential Election.}
\item{MACRON}{Number of votes gained at a national level for Emmanuel Macron in the
first-round of 2017 Presidential Election.}
\item{LE_PEN}{Number of votes gained at a national level for Marine Le Pen in the
first-round of 2017 Presidential Election.}
\item{FILLON}{Number of votes gained at a national level for Francois Fillon in the
first-round of 2017 Presidential Election.}
\item{MELENCHON}{Number of votes gained at a national level for Jean-Luc Melenchon in the
first-round of 2017 Presidential Election.}
\item{HAMON}{Number of votes gained at a national level for Benoit Hamon in the
first-round of 2017 Presidential Election.}
\item{DUPONT.AIGNAN}{Number of votes gained at a national level for Nicolas Dupont-Aignan in the
first-round of 2017 Presidential Election.}
\item{OTHERS}{Number of votes gained at a national level for the rest of candidates in the
first-round of 2017 Presidential Election.}
\item{ABSTENTION2}{Number of people abstaining (NonVoters) in the second-round of 2017 Presidential Election.}
\item{BLANK_NULL2}{Number of people voting either blank or null in the second-round of 2017 Presidential Election.}
\item{MACRON2}{Number of votes gained at a national level for Emmanuel Macron in the
second-round of 2017 Presidential Election}
\item{LE_PEN2}{Number of votes gained at a national level for Marine Le Pen in the
second-round of 2017 Presidential Election}
}}

\source{
Own elaboration from data available in \url{https://www.conseil-constitutionnel.fr/}, retrieved 3 March 2020.
}

\keyword{datasets}
